fix phonebook

+++ cpp00/ex01/PhoneBook.hpp
@@ -6,7 +6,7 @@
 /*   By: lgillard <mirsella@protonmail.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/21 16:12:56 by lgillard          #+#    #+#             */
-/*   Updated: 2023/03/23 22:29:07 by mirsella         ###   ########.fr       */
+/*   Updated: 2023/03/31 12:00:30 by lgillard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,13 +21,13 @@ class PhoneBook
 	{
 	private:
 		Contact _contacts[8];
-		int		 _nbContacts;
+		size_t	_nbContacts;
 	public:
 		PhoneBook();
-		int		getNbContacts();
+		size_t	getNbContacts();
 		void	printPhoneBook();
 		int		newContact();
-		Contact	getContact(int index);
+		Contact	getContact(size_t index);
 };
 
 #endif
